requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
networkx>=2.6.0
sqlite3
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
typer>=0.4.0
rich>=12.0.0
loguru>=0.6.0
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
